authorization added

+++ vagrant/catalog/routes/departments.py
@@ -17,30 +17,36 @@ def departmentList():
     listOfDeparments = controller.getAllDepartments()
     return render_template('department.html', departments=listOfDeparments)
 
-
 @routes.route('/Deparment/new/', methods=['GET','POST'])
 def newDepartment():
+    currentUser = login_session['user_id']
     # Check to see if user is logged in to get to page
     if 'username' not in login_session:
         return redirect('/login')
     if request.method =='POST':
         if request.form['name'] and request.form['description']:
-            editName = request.form['name']
-            editDesc = request.form['description']
+            addName = request.form['name']
+            addDesc = request.form['description']
+            addUser = currentUser.id
             # Add department to controller
-            controller.addNewDepartment(editName,editDesc)
+            controller.addNewDepartment(addName,addDesc,addUser)
             flash("Department Added Successfully!")
         return redirect(url_for('routes.departmentList'))
     else:
         return render_template('newdepartment.html')
-            
-    return "page to add Deparment"
 
 @routes.route('/Deparment/<int:department_id>/edit/', methods=['GET','POST'])
 def editDepartment(department_id):
+    getDepartmentUser = controller.getDepartmentUser(department_id)
+    currentUser = login_session['user_id']
     # Check to see if user is logged in to get to page
     if 'username' not in login_session:
         return redirect('/login')
+    if getDepartmentUser != currentUser.id:
+        msg ="""You are not authorized to edit this Department. Please 
+                create your own Department to edit!"""
+        flash(msg)
+        return redirect(url_for('routes.departmentList'))
     if request.method =='POST':
         if request.form['name']:
             editName = request.form['name']
@@ -59,9 +65,16 @@ def editDepartment(department_id):
 
 @routes.route('/Deparment/<int:department_id>/delete/', methods=['GET','POST'])
 def deleteDepartment(department_id):
+    getDepartmentUser = controller.getDepartmentUser(department_id)
+    currentUser = login_session['user_id']
     # Check to see if user is logged in to get to page
     if 'username' not in login_session:
         return redirect('/login')
+    if getDepartmentUser != currentUser.id:
+        msg = """You are not authorized to delete this Department. Please 
+                 create your own Department to delete!"""
+        flash(msg)
+        return redirect(url_for('routes.departmentList'))
     if request.method =='POST':
         # Delete department to controller
         controller.deleteDepartment(department_id)
